pyproject.toml
# pyproject.toml

# Note: Many fields are autogenerated by 'WIPACrepo/wipac-dev-py-setup-action'.
#       See this repo's GitHub Actions workflow(s) (.github/workflows/*.yaml)
#       and https://github.com/WIPACrepo/wipac-dev-py-setup-action for details.

[project]
dependencies = [
    'cachetools',
    'jsonschema',
    'openapi-core',
    'pymongo',
    'wipac-dev-tools',
    'wipac-rest-tools',
]
dynamic = ["version"] # do not edit — autogenerated by wipac-dev-py-setup-action
name = "wms" # do not edit — autogenerated by wipac-dev-py-setup-action
requires-python = ">=3.12, <3.14" # do not edit — autogenerated by wipac-dev-py-setup-action
keywords = [
    'EWMS',
    'task',
    'Workflow Management Service',
    'WIPAC',
    'IceCube',
    'Observation Management Service',
] # do not edit — autogenerated by wipac-dev-py-setup-action

[project.optional-dependencies]
tests = [
    'flask',
    'nest-asyncio',
    'pytest',
    'pytest-asyncio',
    'pytest-mock',
    'pytest-reverse',
]
examples = [
    'oms-mqclient',
]
mypy = [
    'flask',
    'nest-asyncio',
    'oms-mqclient',
    'pytest',
    'pytest-asyncio',
    'pytest-mock',
    'pytest-reverse',
] # do not edit — autogenerated by wipac-dev-py-setup-action

[build-system] # do not edit — autogenerated by wipac-dev-py-setup-action
requires = ["setuptools>=78.1", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools] # do not edit — autogenerated by wipac-dev-py-setup-action
packages = [
    'wms',
    'wms.database',
    'wms.rest_handlers',
    'wms.schema',
] # do not edit — autogenerated by wipac-dev-py-setup-action

[tool.setuptools.package-data]
"*" = ["py.typed"] # do not edit — autogenerated by wipac-dev-py-setup-action

[tool.setuptools_scm] # do not edit — autogenerated by wipac-dev-py-setup-action
fallback_version = "CANNOT_BUILD_WITHOUT_GIT_DIR" # used when SCM metadata is missing (no '.git/') — if '.git/' exists but does not contain tags, a commit-based version (like '0.1.dev1+g79c9b4b') will be used instead.
